Use size_t when ranging over sizes
These address sign-compare warnings in test and benchmarks.

+++ benchmarks/db_bench_sqlite3.cc
@@ -147,11 +147,11 @@ class RandomGenerator {
 };
 
 static Slice TrimSpace(Slice s) {
-  int start = 0;
+  size_t start = 0;
   while (start < s.size() && isspace(s[start])) {
     start++;
   }
-  int limit = s.size();
+  size_t limit = s.size();
   while (limit > start && isspace(s[limit - 1])) {
     limit--;
   }
@@ -323,7 +323,7 @@ class Benchmark {
     Env::Default()->GetTestDirectory(&test_dir);
     Env::Default()->GetChildren(test_dir, &files);
     if (!FLAGS_use_existing_db) {
-      for (int i = 0; i < files.size(); i++) {
+      for (size_t i = 0; i < files.size(); i++) {
         if (Slice(files[i]).starts_with("dbbench_sqlite3")) {
           std::string file_name(test_dir);
           file_name += "/";
